requirements.txt
requests==2.28.2
python-dotenv==0.9.1
python-telegram-bot==13.15